workaround linkrot
The MISB ST 0601 link was broken. The .pdf file hosting is now done by AWS and linking to it directly by file without the AWS arguments in the URL returns an "access denied" error (predictable). Linking to it with the link I have (with dates and access keys in it) won't work for others on other days, so instead I linked to the NGA page on the standard. It's one extra click to get to the .pdf now.